Move functions to init
