+++ pandas/tests/io/conftest.py
@@ -8,7 +8,7 @@ import pytest
 
 import pandas._testing as tm
 
-from pandas.io.parsers.readers import read_csv
+from pandas.io.parsers import read_csv
 
 
 @pytest.fixture
