+++ pandas/tests/io/excel/conftest.py
@@ -4,7 +4,7 @@ import pandas.util._test_decorators as td
 
 import pandas._testing as tm
 
-from pandas.io.parsers.readers import read_csv
+from pandas.io.parsers import read_csv
 
 
 @pytest.fixture
