+++ pandas/tests/io/parser/test_unsupported.py
@@ -14,8 +14,8 @@ from pandas.errors import ParserError
 
 import pandas._testing as tm
 
+from pandas.io.parsers import read_csv
 import pandas.io.parsers.readers as parsers
-from pandas.io.parsers.readers import read_csv
 
 
 @pytest.fixture(params=["python", "python-fwf"], ids=lambda val: val)
